Style: Moved Stix into its own module
