+++ Code/step_comparison.py
@@ -4,10 +4,12 @@ import datetime
 import matplotlib as mpl
 import matplotlib.pyplot as plt
 
-import stix_handler
+import epd.data_helper
+import epd.loader
+import stix
 import connectivity_tool
 import plots
-import epd_handler
+import epd
 import misc
 import math
 import config
@@ -35,10 +37,10 @@ end_date = config.END_DATE
 delta = 10      # radius of connection points that get accepted (degrees)
 
 # read STIX flare list and extract coordinates of the origin
-stix_flares = stix_handler.read_list()
+stix_flares = stix.read_list()
 
 # get range of flares that are within the defined timeframe
-flare_start_id, flare_end_id = stix_handler.flares_range(start_date, end_date, stix_flares['peak_UTC'])
+flare_start_id, flare_end_id = stix.flares_range(start_date, end_date, stix_flares['peak_UTC'])
 
 # returns a list of candidates the MCT (Magnetic Connectivity Tool) expects the Solar Orbiter to be connected with
 # flare_distances is currently not used for anything (has been added as it might yield interesting data)
@@ -55,7 +57,7 @@ sensor = 'step' # ['het', 'ept', 'step']
 
 # columns of step data
 # ['DELTA_EPOCH', 'Integral_Avg_Flux_0-47', 'Integral_Avg_Uncertainty_0-47', 'Magnet_Avg_Flux_0-47', 'Magnet_Avg_Uncertainty_0-47', 'QUALITY_BITMASK', 'QUALITY_FLAG', 'SMALL_PIXELS_FLAG']
-df_step = epd_handler.load_pickles(sensor, start_date, end_date)
+df_step = epd.loader.load_pickles(sensor, start_date, end_date)
 
 # remove flag columns as currently not used
 drop_columns = ['DELTA_EPOCH', 'QUALITY_BITMASK', 'QUALITY_FLAG']
@@ -91,7 +93,7 @@ for i in range(length):
     df_step_electron[electron_cols[i]] = pd.Series(df_integral[integral_col].to_numpy() - df_magnet[magnet_col].to_numpy(), index = df_step_electron.index)  
 
 # compute running averade and standard deviation
-running_mean, running_std = epd_handler.running_average(df_step_electron)
+running_mean, running_std = epd.data_helper.running_average(df_step_electron)
 
 print("Running averages computed...")
 
@@ -124,7 +126,7 @@ for i in df_offset_step.columns:
 
 # try to find events in data
 sigma_factor = 3.5
-events = epd_handler.find_event(df_offset_step, running_mean, running_std, sigma_factor)
+events = epd.data_helper.find_event(df_offset_step, running_mean, running_std, sigma_factor)
 
 print("Events found...")
